feat: Expose createPermissionMiddleware via messenger
