+++ packages/permission-controller/src/PermissionController.ts
@@ -23,7 +23,7 @@ import type {
 } from '@metamask/messenger';
 import { JsonRpcError } from '@metamask/rpc-errors';
 import { hasProperty } from '@metamask/utils';
-import type { Json, Mutable } from '@metamask/utils';
+import type { Json, JsonRpcRequest, Mutable, PendingJsonRpcResponse } from '@metamask/utils';
 import deepFreeze from 'deep-freeze-strict';
 import { castDraft, produce as immerProduce } from 'immer';
 import type { Draft } from 'immer';
@@ -93,10 +93,10 @@ import {
   hasSpecificationType,
   PermissionType,
 } from './Permission';
-import { getPermissionMiddlewareFactory } from './permission-middleware';
 import type { PermissionControllerMethodActions } from './PermissionController-method-action-types';
 import type { SubjectMetadataControllerGetSubjectMetadataAction } from './SubjectMetadataController-method-action-types';
 import { collectUniqueAndPairedCaveats, MethodNames } from './utils';
+import { AsyncJsonRpcEngineNextCallback, createAsyncMiddleware, JsonRpcMiddleware } from '@metamask/json-rpc-engine';
 
 /**
  * Flags for controlling the validation behavior of certain internal methods.
@@ -190,6 +190,7 @@ const MESSENGER_EXPOSED_METHODS = [
   'revokePermissions',
   'updateCaveat',
   'getCaveat',
+  'createPermissionMiddleware',
 ] as const;
 
 /**
@@ -228,10 +229,10 @@ export type PermissionControllerSubjects<
  */
 export type PermissionControllerState<Permission> =
   Permission extends PermissionConstraint
-    ? {
-        subjects: PermissionControllerSubjects<Permission>;
-      }
-    : never;
+  ? {
+    subjects: PermissionControllerSubjects<Permission>;
+  }
+  : never;
 
 /**
  * Get the state metadata of the {@link PermissionController}.
@@ -515,20 +516,20 @@ export type CaveatMutator<TargetCaveat extends CaveatConstraint> = (
 
 type CaveatMutatorResult =
   | Readonly<{
-      operation: CaveatMutatorOperation.UpdateValue;
-      value: CaveatConstraint['value'];
-    }>
+    operation: CaveatMutatorOperation.UpdateValue;
+    value: CaveatConstraint['value'];
+  }>
   | Readonly<{
-      operation: Exclude<
-        CaveatMutatorOperation,
-        CaveatMutatorOperation.UpdateValue
-      >;
-    }>;
+    operation: Exclude<
+      CaveatMutatorOperation,
+      CaveatMutatorOperation.UpdateValue
+    >;
+  }>;
 
 type MergeCaveatResult<CaveatType extends CaveatConstraint | undefined> =
   CaveatType extends undefined
-    ? [CaveatConstraint, CaveatConstraint['value']]
-    : [CaveatConstraint, CaveatConstraint['value']] | [];
+  ? [CaveatConstraint, CaveatConstraint['value']]
+  : [CaveatConstraint, CaveatConstraint['value']] | [];
 
 /**
  * Extracts the permission(s) specified by the given permission and caveat
@@ -545,11 +546,11 @@ export type ExtractPermission<
   ControllerCaveatSpecification extends CaveatSpecificationConstraint,
 > =
   ControllerPermissionSpecification extends ValidPermissionSpecification<ControllerPermissionSpecification>
-    ? ValidPermission<
-        ControllerPermissionSpecification['targetName'],
-        ExtractCaveats<ControllerCaveatSpecification>
-      >
-    : never;
+  ? ValidPermission<
+    ControllerPermissionSpecification['targetName'],
+    ExtractCaveats<ControllerCaveatSpecification>
+  >
+  : never;
 
 /**
  * Extracts the restricted method permission(s) specified by the given
@@ -632,9 +633,9 @@ export type PermissionControllerOptions<
  */
 export class PermissionController<
   ControllerPermissionSpecification extends PermissionSpecificationConstraint =
-    PermissionSpecificationConstraint,
+  PermissionSpecificationConstraint,
   ControllerCaveatSpecification extends CaveatSpecificationConstraint =
-    CaveatSpecificationConstraint,
+  CaveatSpecificationConstraint,
 > extends BaseController<
   typeof controllerName,
   PermissionControllerState<
@@ -663,18 +664,6 @@ export class PermissionController<
   public get unrestrictedMethods(): ReadonlySet<string> {
     return this.#unrestrictedMethods;
   }
-
-  /**
-   * Returns a `json-rpc-engine` middleware function factory, so that the rules
-   * described by the state of this controller can be applied to incoming
-   * JSON-RPC requests.
-   *
-   * The middleware **must** be added in the correct place in the middleware
-   * stack in order for it to work. See the README for an example.
-   */
-  public createPermissionMiddleware: ReturnType<
-    typeof getPermissionMiddlewareFactory
-  >;
 
   /**
    * Constructs the PermissionController.
@@ -744,14 +733,6 @@ export class PermissionController<
       this,
       MESSENGER_EXPOSED_METHODS,
     );
-
-    this.createPermissionMiddleware = getPermissionMiddlewareFactory({
-      executeRestrictedMethod: this.#executeRestrictedMethod.bind(this),
-      getRestrictedMethod: this.getRestrictedMethod.bind(this),
-      isUnrestrictedMethod: this.unrestrictedMethods.has.bind(
-        this.unrestrictedMethods,
-      ),
-    });
   }
 
   /**
@@ -844,7 +825,7 @@ export class PermissionController<
 
             const specification =
               caveatSpecifications[
-                caveatType as ControllerCaveatSpecification['type']
+              caveatType as ControllerCaveatSpecification['type']
               ];
             const isRestrictedMethodCaveat =
               isRestrictedMethodCaveatSpecification(specification);
@@ -904,13 +885,13 @@ export class PermissionController<
     const failureError =
       permissionType === PermissionType.RestrictedMethod
         ? methodNotFound(
-            targetName,
-            requestingOrigin ? { origin: requestingOrigin } : undefined,
-          )
+          targetName,
+          requestingOrigin ? { origin: requestingOrigin } : undefined,
+        )
         : new EndowmentPermissionDoesNotExistError(
-            targetName,
-            requestingOrigin,
-          );
+          targetName,
+          requestingOrigin,
+        );
 
     if (!this.#targetExists(targetName)) {
       throw failureError;
@@ -923,6 +904,72 @@ export class PermissionController<
 
     return specification;
   }
+
+  /**
+   * Creates a permission middleware function. Like any {@link JsonRpcEngine}
+   * middleware, each middleware will only receive requests from a particular
+   * subject / origin.
+   *
+   * The middlewares returned will pass through requests for
+   * unrestricted methods, and attempt to execute restricted methods. If a method
+   * is neither restricted nor unrestricted, a "method not found" error will be
+   * returned.
+   * If a method is restricted, the middleware will first attempt to retrieve the
+   * subject's permission for that method. If the permission is found, the method
+   * will be executed. Otherwise, an "unauthorized" error will be returned.
+   *
+   * The middleware **must** be added in the correct place in the middleware
+   * stack in order for it to work. See the README for an example.
+   *
+   * @param subject The permission subject.
+   * @returns A `json-rpc-engine` middleware.
+   */
+  public createPermissionMiddleware(
+    subject: PermissionSubjectMetadata,
+  ): JsonRpcMiddleware<RestrictedMethodParameters, Json> {
+    const { origin } = subject;
+    if (typeof origin !== 'string' || !origin) {
+      throw new Error('The subject "origin" must be a non-empty string.');
+    }
+
+    const permissionsMiddleware = async (
+      req: JsonRpcRequest<RestrictedMethodParameters>,
+      res: PendingJsonRpcResponse,
+      next: AsyncJsonRpcEngineNextCallback,
+    ): Promise<void> => {
+      const { method, params } = req;
+
+      // Skip registered unrestricted methods.
+      if (this.#unrestrictedMethods.has(method)) {
+        return next();
+      }
+
+      // This will throw if no restricted method implementation is found.
+      const methodImplementation = this.getRestrictedMethod(method, origin);
+
+      // This will throw if the permission does not exist.
+      const result = await this.#executeRestrictedMethod(
+        methodImplementation,
+        subject,
+        method,
+        params,
+      );
+
+      if (result === undefined) {
+        res.error = internalError(
+          `Request for method "${req.method}" returned undefined result.`,
+          { request: req },
+        );
+        return undefined;
+      }
+
+      res.result = result;
+      return undefined;
+    };
+
+    return createAsyncMiddleware(permissionsMiddleware);
+  };
+
 
   /**
    * Gets the implementation of the specified restricted method.
@@ -989,8 +1036,8 @@ export class PermissionController<
     origin: OriginString,
   ):
     | SubjectPermissions<
-        ValidPermission<string, ExtractCaveats<ControllerCaveatSpecification>>
-      >
+      ValidPermission<string, ExtractCaveats<ControllerCaveatSpecification>>
+    >
     | undefined {
     return this.state.subjects[origin]?.permissions;
   }
@@ -1170,7 +1217,7 @@ export class PermissionController<
       ControllerCaveatSpecification
     >['parentCapability'],
     CaveatType extends
-      ExtractAllowedCaveatTypes<ControllerPermissionSpecification>,
+    ExtractAllowedCaveatTypes<ControllerPermissionSpecification>,
   >(origin: OriginString, target: TargetName, caveatType: CaveatType): boolean {
     return Boolean(this.getCaveat(origin, target, caveatType));
   }
@@ -1196,7 +1243,7 @@ export class PermissionController<
       ControllerCaveatSpecification
     >['parentCapability'],
     CaveatType extends
-      ExtractAllowedCaveatTypes<ControllerPermissionSpecification>,
+    ExtractAllowedCaveatTypes<ControllerPermissionSpecification>,
   >(
     origin: OriginString,
     target: TargetName,
@@ -1237,7 +1284,7 @@ export class PermissionController<
       ControllerCaveatSpecification
     >['parentCapability'],
     CaveatType extends
-      ExtractAllowedCaveatTypes<ControllerPermissionSpecification>,
+    ExtractAllowedCaveatTypes<ControllerPermissionSpecification>,
   >(
     origin: OriginString,
     target: TargetName,
@@ -1275,7 +1322,7 @@ export class PermissionController<
       ControllerCaveatSpecification
     >['parentCapability'],
     CaveatType extends
-      ExtractAllowedCaveatTypes<ControllerPermissionSpecification>,
+    ExtractAllowedCaveatTypes<ControllerPermissionSpecification>,
     CaveatValue extends ExtractCaveatValue<
       ControllerCaveatSpecification,
       CaveatType
@@ -1317,7 +1364,7 @@ export class PermissionController<
       ControllerCaveatSpecification
     >['parentCapability'],
     CaveatType extends
-      ExtractAllowedCaveatTypes<ControllerPermissionSpecification>,
+    ExtractAllowedCaveatTypes<ControllerPermissionSpecification>,
   >(
     origin: OriginString,
     target: TargetName,
@@ -1489,7 +1536,7 @@ export class PermissionController<
   removeCaveat<
     TargetName extends ControllerPermissionSpecification['targetName'],
     CaveatType extends
-      ExtractAllowedCaveatTypes<ControllerPermissionSpecification>,
+    ExtractAllowedCaveatTypes<ControllerPermissionSpecification>,
   >(origin: OriginString, target: TargetName, caveatType: CaveatType): void {
     this.update((draftState) => {
       const permission = draftState.subjects[origin]?.permissions[target];
@@ -1721,8 +1768,8 @@ export class PermissionController<
     const permissions = (
       preserveExistingPermissions
         ? {
-            ...this.getPermissions(origin),
-          }
+          ...this.getPermissions(origin),
+        }
         : {}
     ) as SubjectPermissions<
       ExtractPermission<
@@ -2093,16 +2140,16 @@ export class PermissionController<
     } = {},
   ): Promise<
     | [
-        Partial<
-          SubjectPermissions<
-            ExtractPermission<
-              ControllerPermissionSpecification,
-              ControllerCaveatSpecification
-            >
+      Partial<
+        SubjectPermissions<
+          ExtractPermission<
+            ControllerPermissionSpecification,
+            ControllerCaveatSpecification
           >
-        >,
-        ApprovedPermissionsMetadata,
-      ]
+        >
+      >,
+      ApprovedPermissionsMetadata,
+    ]
     | []
   > {
     const { origin } = subject;
@@ -2244,11 +2291,11 @@ export class PermissionController<
     incrementalRequestedPermissions: RequestedPermissions,
   ):
     | [
-        SubjectPermissions<
-          ValidPermission<string, ExtractCaveats<ControllerCaveatSpecification>>
-        >,
-        PermissionDiffMap<string, CaveatConstraint>,
-      ]
+      SubjectPermissions<
+        ValidPermission<string, ExtractCaveats<ControllerCaveatSpecification>>
+      >,
+      PermissionDiffMap<string, CaveatConstraint>,
+    ]
     | [] {
     const permissionDiffMap: PermissionDiffMap<string, CaveatConstraint> = {};
 
@@ -2388,12 +2435,12 @@ export class PermissionController<
 
     return newValue !== undefined && diff !== undefined
       ? [
-          {
-            type: rightCaveat.type,
-            value: newValue,
-          },
-          diff,
-        ]
+        {
+          type: rightCaveat.type,
+          value: newValue,
+        },
+        diff,
+      ]
       : ([] as MergeCaveatResult<LeftCaveat>);
   }
 
@@ -2554,9 +2601,9 @@ export class PermissionController<
 
       throw reasons.length > 1
         ? internalError(
-            'Multiple errors occurred during side-effects execution',
-            { errors: reasons },
-          )
+          'Multiple errors occurred during side-effects execution',
+          { errors: reasons },
+        )
         : reasons[0];
     }
